[db] Move custom functions to GIN file
internal/db/user.go file contains no custom function definitions now.
